move asset.get to graph.get
